refactor(mm-next): add `<Head>` in story, story/amp page
